4045: Move BottomTabNavigator and url helper
