+++ native/src/routes/TuNewsDetail.tsx
@@ -11,7 +11,7 @@ import useHeader from '../hooks/useHeader'
 import { CityContentData } from '../hooks/useLoadCityContent'
 import useLoadTuNewsElement from '../hooks/useLoadTuNewsElement'
 import useSetRouteTitle from '../hooks/useSetRouteTitle'
-import urlFromRouteInformation from '../navigation/url'
+import urlFromRouteInformation from '../utils/url'
 import LoadingErrorHandler from './LoadingErrorHandler'
 
 const HeaderImageWrapper = styled.View`
